renamed a folder and modified where it's imported
